==> src/01_Data_Cleaning_myusabank.ipynb <==
# Import Libraries and Load the Data

## First, import the necessary libraries and load the dataset into a pandas DataFrame.

import pandas as pd
import numpy as np
import seaborn as sns
import datetime

## Load the dataset and Preview 

df = pd.read_csv(r"C:\Users\shres\Desktop\codekentucky\usabank-project\data\raw\myusabank.csv")
df.head(5)

df.tail(5)

## Find data information

df.info()

## Find  Columns and Rows

df.shape

# Data Cleaning and Preprocessing
 

## Check for null values

df.isnull().sum()
print(df.isnull().sum())

### Drop null values

df.dropna(inplace = True)

### Check for null values after Drop null values if any

#df.isnull().sum()
#print(df.isnull().sum())
df.shape

## Handling Duplicates 
Check for duplicates and remove them if present

### Check duplicates

#df.duplicated().sum()
duplicates = df.duplicated().sum()
print(f"Number of duplicate rows: {duplicates}")

### Remove duplicates

df = df.drop_duplicates()

### Check duplicates if any

df.duplicated().sum()
#duplicates = df.duplicated().sum()
#print(f"Number of duplicate rows: {duplicates}")

## Find all columns name

df.columns

## Check Data Type

df.dtypes

### Check skewness 

df.skew(numeric_only = 'True')

Describe() method returns description of the data in the data DataFrame.

df.describe()

## Data Type Conversion

#### Convert 'Date' column to datetime format.
    Ensure the Date column is in datetime format.

df['Date'] = pd.to_datetime(df['Date'])

## Scaling and Normalization

columns_to_scale = ['Interest_Income', 'Interest_Expense', 'Average_Earning_Assets', 
                    'Net_Income', 'Total_Assets', 'Shareholder_Equity', 
                    'Operating_Expenses', 'Operating_Income', 'Market_Share', 'Stock_Price']

# Remove columns

## Select only the 'Date' and 'Close' columns

# Convert the date column to datetime format (assuming your date column is named 'Date')
df['Date'] = pd.to_datetime(df['Date'])

# Define the date range as strings
start_date = 'YYYY-MM-DD'
end_date = 'YYYY-MM-DD'
# Select only the 'Date' and 'Close' columns
bank_df = df[['Date', 'Stock_Price']]
bank_df.head()

bank_df.tail()

bank_df.dtypes

Creat the myusabank CSV file

bank_df.to_csv('../data/clean/01_myusabank_clean_data.csv')

Clean myusabank data CSV file send to data folder and save.
