==> src/02_Data_Cleaning_S&P 500(SPX).ipynb <==
# Import Libraries and Load the Data
## First, import the necessary libraries and load the dataset into a pandas DataFrame.

import pandas as pd
import numpy as np
import seaborn as sns
import datetime

# Load the dataset and Preview

df = pd.read_csv(r"C:\Users\shres\Desktop\codekentucky\usabank-project\data\raw\S&P 500 (SPX)HistoricalData_1730734525504.csv")
df.head()

df.tail()

## Find data information

df.info()

## Find  Columns and Rows

df.shape

# Data Cleaning and Preprocessing

## Check for null values

df.isnull().sum()
print(df.isnull().sum())

### Check duplicates

duplicates = df.duplicated().sum()
print(f"Number of duplicate rows: {duplicates}")

## Find all columns name

df.columns

## Check Data Type

df.dtypes

## Describe() method returns description of the data in the data DataFrame.

df.describe()

# Filter the dataframe and Display the filtered data

## Convert the date column to datetime format (assuming your date column is named 'Date')

df['Date'] = pd.to_datetime(df['Date'])

## Define the date range as strings

start_date = '2022-01-03'
end_date = '2023-03-23'

## Filter the dataframe

filtered_df = df[(df['Date'] >= start_date) & (df['Date'] <= end_date)]

## Display the filtered data

filtered_df = df[(df['Date'] >= start_date) & (df['Date'] <= end_date)]
filtered_df.head()


filtered_df.tail()

### Check skewness 

# Calculate skewness for numeric columns only
filtered_df.skew(numeric_only=True)

### Describe() method returns description of the data in the data DataFrame after filter the Data

filtered_df.describe()

# Remove columns

## Select only the 'Date' and 'Close' columns

selected_df = filtered_df[['Date', 'Close/Last']]


## Display the first few rows of the selected data

selected_df = filtered_df[['Date', 'Close/Last']]
selected_df.head()

selected_df.tail()

selected_df.dtypes

Creat the S&P500(SPX) CSV file

selected_df.to_csv('../data/clean/02_S&P500(SPX)_clean_data.csv')

Clean S&P500(SPX) data CSV file send to data folder and save.
