==> src/03_Data_Merge_myusabank_S&P500.ipynb <==
# Import Libraries and Load the Data
### First, import the necessary libraries and load the dataset into a pandas DataFrame.

import pandas as pd
import numpy as np
import seaborn as sns
import datetime

# Load the Bank Data

df = pd.read_csv(r"C:\Users\shres\Desktop\codekentucky\usabank-project\data\clean\01_myusabank_clean_data.csv") 
df['Date'] = pd.to_datetime(df['Date'])

Details: Load data for a bank's stock prices from a CSV file. Reads a CSV file (myusabank.csv) containing bank financial data into a DataFrame named df and converts the Date column from a string format to a datetime object. This makes it easier to filter data based on dates later in the code.

#  Load the S&P 500 Data

spx_df = pd.read_csv(r"C:\Users\shres\Desktop\codekentucky\usabank-project\data\clean\02_S&P500(SPX)_clean_data.csv")
spx_df['Date'] = pd.to_datetime(spx_df['Date'])

Details: Load historical S&P 500 data for comparison with the bank's stock prices. Reads another CSV file (S&P 500 historical data) into a DataFrame named spx_df. It converts the Date column to a datetime object for easy filtering by date range.

# Define the Date Range for Analysis

start_date = '2022-01-03'
end_date = '2023-03-23'

Details: Set the start and end dates for the data range of interest and it assigns two specific dates that act as filters, allowing you to focus on data from early 2022 through early 2023.

# Filter Both DataFrames Based on the Date Range

bank_df = df[(df['Date'] >= start_date) & (df['Date'] <= end_date)][['Date', 'Stock_Price']]
spx_df_filtered = spx_df[(spx_df['Date'] >= start_date) & (spx_df['Date'] <= end_date)][['Date', 'Close/Last']]

Purpose: Isolate the relevant rows in both the bank and S&P 500 DataFrames based on the specified date range.
Details:
    Filters the df DataFrame (bank data) for rows where the Date falls within the date range, and selects only the Date and Stock_Price columns.
    Similarly, filters the spx_df DataFrame (S&P 500 data) for dates within the range and selects the Date and Close/Last columns.

# Rename Columns for Clarity

bank_df = bank_df.rename(columns={'Stock_Price': 'Bank_Stock_Price'})
spx_df_filtered = spx_df_filtered.rename(columns={'Close/Last': 'S&P500'})

bank_df = bank_df.rename(columns={'date': 'Date'})
spx_df_filtered = spx_df_filtered.rename(columns={'date': 'Date'})

Details: Make column names clearer in both DataFrames and renames Stock_Price in bank_df to Bank_Stock_Price, helping distinguish it from the S&P 500 data, renames Close/Last in spx_df_filtered to S&P500, making it clear that this column contains S&P 500 closing prices.

# Merge the DataFrames on the 'Date' Column

merged_df = pd.merge(bank_df, spx_df_filtered, how = 'inner', on = ['Date'])

merged_df['Date'] = pd.to_datetime(merged_df['Date']) 

print(merged_df.columns)

Purpose: Combine the bank and S&P 500 data into a single DataFrame based on matching dates.
Details:
    Uses an inner join to merge bank_df and spx_df_filtered on the Date column.
    The how='inner' parameter ensures that only rows with dates present in both data sets are included.

# Display the First Few Rows of the Merged Data

merged_df.head()

Details: Inspect the initial rows of the merged data and head() displays the first few rows, allowing you to verify that the merge worked correctly and that the columns (Date, Bank_Stock_Price, and S&P500) contain the expected data.

merged_df.tail()

Details: Inspect the initial rows of the merged data and tail() displays the first few rows, allowing you to verify that the merge worked correctly and that the columns (Date, Bank_Stock_Price, and S&P500) contain the expected data.

### Check the  description

merged_df.describe()

Details: Inspect the  merged data  description displays the count, mean, min max and std  to verify that the merge worked correctly and that the columns (Date, Bank_Stock_Price, and S&P500) contain the expected data.

### Check Data Types

merged_df.dtypes

### Creat the Merge CSV File 

merged_df.to_csv('../data/clean/03_data_merged_myusabank_S&P500_data.csv', index=False) #Unnamed: 0 show the colume do not show the file

CSV file send to data folder and save.
